chore: improve logger
